Explicit signs for < comparisons

+++ tests/end_to_end/float-conversions/irix-o2-out.c
@@ -9,14 +9,14 @@ void *test(void)
     D_410240 = (u32) D_410238;
     temp_f18 = (f64) D_410240;
     phi_f18 = temp_f18;
-    if (D_410240 < 0)
+    if ((s32) D_410240 < 0)
     {
         phi_f18 = temp_f18 + 4294967296.0;
     }
     D_410238 = phi_f18;
     temp_f8 = (f32) D_410240;
     phi_f8 = temp_f8;
-    if (D_410240 < 0)
+    if ((s32) D_410240 < 0)
     {
         phi_f8 = temp_f8 + 4294967296.0f;
     }
